[BUGFIX] Use `get_called_class` instead of `__CLASS__`
Summary:
To get the class properties `get_called_class` should
be used instead of `__CLASS__` as `__CLASS__` returns the class
name where the method / trait is implemented and not the class
of the current instance.
That means that the export function does not export properties
of the inherited class but only those of the class where the
trait is implemented.

Test Plan: Verify and execute unit tests.

Reviewers: claus.due

Reviewed By: claus.due

Differential Revision: https://phabricator.dkd.de/D882