[FEATURE] Allow property name as getter but fail if multiple access methods found
Summary:
In some cases the property name can also be the getter method name.
This happens for example for properties that have the name `isFoo`. The
getter could then be named `isFoo()` and the setter `setIsFoo($foo)`.

To prevent unexpected behavior a exception is now thrown if there are
multiple getters or setter found for a property.

Test Plan: Check and run unit tests

Reviewers: claus.due

Reviewed By: claus.due

Differential Revision: https://phabricator.dkd.de/D887